on focus change should be a method only for text input
