+++ src/core/tests/widgets/test_base.py
@@ -13,14 +13,9 @@ class WidgetTests(TestCase):
         self.id = 'widget_id'
         self.style = Pack(padding=666)
 
-        self.on_gain_focus = mock.Mock()
-        self.on_lose_focus = mock.Mock()
-
         self.widget = toga.Widget(
             id=self.id,
             style=self.style,
-            on_gain_focus=self.on_gain_focus,
-            on_lose_focus=self.on_lose_focus,
             factory=toga_dummy.factory
         )
 
@@ -64,9 +59,3 @@ class WidgetTests(TestCase):
         self.widget._children = []
         self.widget.add(child1, child2)
         self.assertEqual(self.widget.children, [child1, child2])
-
-    def test_on_gain_focus(self):
-        self.assertEqual(self.widget.on_gain_focus, self.on_gain_focus)
-
-    def test_on_lose_focus(self):
-        self.assertEqual(self.widget.on_lose_focus, self.on_lose_focus)
