+++ src/core/toga/widgets/button.py
@@ -12,32 +12,17 @@ class Button(Widget):
         style (:obj:`Style`): An optional style object. If no style is provided then
             a new one will be created for the widget.
         on_press (:obj:`callable`): Function to execute when pressed.
-        on_gain_focus (:obj:`callable`): Function to execute when get focused.
-        on_lose_focus (:obj:`callable`): Function to execute when lose focus.
-        enabled (bool): Whether or not interaction with the button is possible, defaults to `True`.
+        enabled (bool): Whether or not interaction with the button is possible,
+            defaults to `True`.
         factory (:obj:`module`): A python module that is capable to return a
-            implementation of this class with the same name. (optional & normally not needed)
+            implementation of this class with the same name. (optional & normally not
+            needed)
     """
 
     def __init__(
-            self,
-            label,
-            id=None,
-            style=None,
-            on_press=None,
-            on_gain_focus=None,
-            on_lose_focus=None,
-            enabled=True,
-            factory=None
+            self, label, id=None, style=None, on_press=None, enabled=True, factory=None
     ):
-        super().__init__(
-            id=id,
-            style=style,
-            enabled=enabled,
-            on_gain_focus=on_gain_focus,
-            on_lose_focus=on_lose_focus,
-            factory=factory,
-        )
+        super().__init__(id=id, style=style, enabled=enabled, factory=factory)
 
         # Create a platform specific implementation of a Button
         self._impl = self.factory.Button(interface=self)
@@ -45,8 +30,6 @@ class Button(Widget):
         # Set all the properties
         self.label = label
         self.on_press = on_press
-        self.on_gain_focus = on_gain_focus
-        self.on_lose_focus = on_lose_focus
         self.enabled = enabled
 
     @property
